Add files via upload
The application allows to either create new orders, retrieve invoices or show the opening times. When it opens it prints the item list of all available items in the store. When a new order is selected, the user can add and remove items from the shopping basket by typing photo ID. When the user completes the order, it will ask to fill in customer details like name, email address, phone number and address and generate a customer ID. When the application has received all the right information it will generate an order ID and calculate the order total, order date, production time and pick up date. Also, it will save all this data as an invoice in a JSON file. When reviewing orders is selected, the user can choose between searching an order through either order ID, customer ID or customer name or choose to show all orders. When the user chooses to search it will allow to type the chosen ID or name and will show the order(s) associated with that.